add video in apod
